Preare draft for .env bug
