+++ server/package.json
@@ -9,15 +9,15 @@
   },
   "dependencies": {
     "cookie-parser": "^1.4.5",
-    "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "http-errors": "~1.8.0",
     "morgan": "^1.10.0",
     "nodemon": "^2.0.6"
   },
   "devDependencies": {
-    "chai": "^4.2.0",
+    "chai": "^4.3.0",
     "chai-http": "^4.3.0",
+    "dotenv": "^8.2.0",
     "mocha": "^8.2.1"
   }
 }
